Feat: Cart, Favorite API connect postgres

+++ src/modules/favorite/favorite.service.ts
@@ -9,6 +9,7 @@ import { FavoriteProductDto } from './dto/favorite-request.dto';
 import { FavoriteEntity } from './entities/favorite.entity';
 import { messages } from '@/common/resources';
 import { FavoriteRepositoryInterface } from './interface/favorite-repository.interface';
+import { ProductsRepositoryInterface } from '../products/interfaces/product-repository.interface';
 
 @Injectable()
 export class FavoriteService {
@@ -31,6 +32,7 @@ export class FavoriteService {
   async upload(id: number, { product_id }: FavoriteProductDto) {
     try {
       const favorite = await this.saveFavorite(id, product_id);
+      console.log('favorite', favorite);
       return favorite;
     } catch (err) {
       this.logger.error(err);
@@ -50,14 +52,14 @@ export class FavoriteService {
     return `This action updates a #${id} favorite`;
   }
 
-  async saveFavorite(user_id: number, product_id: number) {
+  async saveFavorite(userId: number, productId: number) {
     try {
       const favorite = FavoriteEntity.create({
-        user_id,
-        product_id,
+        user_id: userId,
+        product_id: productId,
       });
       await this.repo.create(favorite);
-      return favorite.toFavorite;
+      return favorite.toFavorite();
     } catch (err) {
       this.logger.error(err);
       throw new UnprocessableEntityException(
@@ -68,6 +70,12 @@ export class FavoriteService {
 
   async removeFavorite(id: number, product_id: number) {
     try {
+      const favorite = await this.repo.findfavoriteByUserId(id, product_id);
+      if (!favorite) {
+        throw new UnprocessableEntityException(
+          messages.FAVORITE_NOT_FOUND_EXCEPTION,
+        );
+      }
       await this.repo.deleteFavoriteByUserId(id, product_id);
     } catch (err) {
       this.logger.error(err);
